requirements.txt
aiogram==3.0.0
asyncpg==0.29.0
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.5.0
sqlparse==0.4.4
psycopg2-binary==2.9.9
httpx==0.25.1
pydantic-settings==2.1.0